+++ app/controllers/application_controller.rb
@@ -3,20 +3,7 @@ class ApplicationController < ActionController::Base
   class Forbidden < ActionController::ActionControllerError; end
   class IpAddressRejected < ActionController::ActionControllerError; end
 
-  rescue_from StandardError, with: :rescue500
-  rescue_from ApplicationController::Forbidden, with: :rescue403
-  rescue_from ApplicationController::IpAddressRejected, with: :rescue403
-
-
-  private def rescue403(e)
-    @exception = e
-    render "errors/forbidden", status: 403
-    # 403は要求されらリソースがwebサイトに存在するがなんらかの理由でアクセス拒否された時
-  end
-
-  private def rescue500(e)
-    render "errors/internal_server_error", status: 500
-    #　500は何らかのエラーが発生してる時
-  end
-  # ActionController::ActionControllerError < StandardError < Exception
+  include ErrorHandlers if Rails.env.production
+  # エラーの表示をプロダクト環境のみ変更するモジュール
+  # concerns/error_handlers参照views/errors参照
 end
